utility class for centering the map
