+++ src/components/sidebar.js
@@ -46,6 +46,4 @@ class Sidebar extends React.Component {
   }
 }
 
-
-
 export default Sidebar;
